requirements.txt
accelerate==1.6.0
einops
gradio
huggingface_hub
insightface
matplotlib
numpy
opencv-python
opencv-python-headless
optimum
optimum_quanto
Pillow
PyYAML
PyYAML
safetensors
seaborn
scikit-image
torch==2.5.1
torchvision==0.20.1
tqdm
transformers==4.45.2
onnxruntime
onnxruntime-gpu
sentencepiece